refactor(avatars): inline safe fetch in favicon action
